Atualização para microserviços

+++ src/main/java/com/letscode/modulobanco811/dtos/ContaRequest.java
@@ -15,5 +15,5 @@ public class ContaRequest {
     private Integer agencia;
     @Enumerated(EnumType.STRING)
     private TipoConta tipoConta;
-    private BigDecimal saldo;
+    private Float saldo;
 }
